+++ src/components/atoms/InputRange/helpers/index.ts
@@ -1,0 +1,25 @@
+interface ITrackPosition {
+  min?: number;
+  max?: number;
+  value: number;
+}
+
+export const calculateTrackPosition = ({ min = 0, max = 100, value }: ITrackPosition) => {
+  if (max < min) {
+    return 0;
+  }
+
+  if (value > max) {
+    return 1;
+  }
+
+  if (value < min) {
+    return 0;
+  }
+
+  if (value === min && value === max) {
+    return 0;
+  }
+
+  return (value - min) / (max - min);
+};